Kth smallest difference
You are given an array of integers. Consider absolute difference between all the pairs of the the elements. You need to find Kth smallest absolute difference. If the size of the array is N then value of K will be less than N and more than or equal to 1.

1
6 2
1 3 4 1 3 8

Output :
0